[Pal] Partially remove PAL_PTR type
Only less controversial places for now, where the replacement type is
quite obvious.

+++ LibOS/shim/include/shim_tcb.h
@@ -12,7 +12,7 @@
 struct shim_context {
     PAL_CONTEXT* regs;
     long syscall_nr;
-    unsigned long tls; /* Used only in clone. */
+    uintptr_t tls; /* Used only in clone. */
 };
 
 typedef struct shim_tcb shim_tcb_t;
